refactor(blockapis): make CachingHttpClient available in src/
We want to be able to use it in `utxo-bin`.

Issue: BG-60311
