+++ modules/blockapis/src/index.ts
@@ -3,6 +3,7 @@ export * from './AddressApi';
 export * from './TransactionApi';
 export * from './UtxoApi';
 export * from './BaseHttpClient';
+export * from './CachingHttpClient';
 
 export * from './impl/BlockchairApi';
 export * from './impl/BlockstreamApi';
